refactor(navigation): structure files
