+++ src/_shell/navigation/desktop/index.ts
@@ -1,1 +1,1 @@
-export * from './nav-menu';
+export * from './desktop-menu';
