--- src/_shell/navigation/desktop/nav-top-menu-item.type.ts
@@ -1,10 +1,0 @@
-import type { MenuAd } from './nav-menu-ad.type';
-import type { NavMenuItem } from './nav-menu-item.type';
-
-export type NavTopMenuItem = {
-  text: string;
-  path?: string;
-  items?: NavMenuItem[];
-  ad?: MenuAd;
-  isCta?: boolean;
-};
