+++ src/_shell/navigation/mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './mobile-navigation';
