+++ src/_shell/navigation/types/entry-menu-item.type.ts
@@ -1,0 +1,10 @@
+import type { MenuAd } from './menu-ad.type';
+import type { MenuItem } from './menu-item.type';
+
+export type MenuEntryItem = {
+  text: string;
+  path?: string;
+  items?: MenuItem[];
+  ad?: MenuAd;
+  isCta?: boolean;
+};
