+++ src/_shell/navigation/types/index.ts
@@ -1,0 +1,4 @@
+export * from './entry-menu-item.type';
+export * from './menu-ad.type';
+export * from './menu-config.type';
+export * from './menu-item.type';
