+++ src/_shell/navigation/types/menu-config.type.tsx
@@ -1,0 +1,5 @@
+import type { MenuEntryItem } from './entry-menu-item.type';
+
+export type MenuConfig = {
+  items: MenuEntryItem[];
+};
