+++ src/_shell/navigation/types/menu-item.type.ts
@@ -1,10 +1,10 @@
 import type { JSXOutput } from '@builder.io/qwik';
 
-export type NavMenuItem = {
+export type MenuItem = {
   text: string;
   path?: string;
   isCategory?: boolean;
   description?: string;
   icon?: JSXOutput;
-  items?: NavMenuItem[];
+  items?: MenuItem[];
 };
